chapter_004/src/main/java/ru/job4j/lambda/Computer.java
package ru.job4j.lambda;

import java.util.LinkedList;
import java.util.List;
import java.util.function.Function;

public class Computer {

    /**
     * Computes function values in a given range
     * @param start range start
     * @param end range end
     * @param func function
     * @return list of computed values
     */
    public List<Double> diapason(int start, int end, Function<Double, Double> func) {
        List<Double> result = new LinkedList<>();
        for (int i = start; i <= end; i++) {
            result.add(func.apply((double) i));
        }
        return result;
    }
}